Added functionality to edit.php
I had to use $_SERVER rather than $_POST['editBtn'] because for some reason it wasn't posting correctly when the button was pushed. I still have the isset in there, and am going to try to figure out why this script doesnt work. In the meantime, $_SERVER['REQUEST_METHOD'] is serving it's purpose.